backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-socketio==5.10.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pgvector==0.2.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Google APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0

# LLM & AI
langchain==0.3.26
langchain-core==0.3.69
langgraph==0.5.3
langchain-openai==0.3.28
langchain-anthropic==0.3.0
langchain-google-genai>=2.0.0
openai>=1.25.0
anthropic>=0.7.7
google-generativeai>=0.3.1

# Data Processing
pydantic==2.5.0
pydantic-settings==2.1.0
numpy==1.25.2
pandas==2.1.3

# Task Scheduling
apscheduler==3.10.4

# Logging & Monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
cryptography==41.0.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
passlib==1.7.4
